chore: fix build issue due to rename folder
